feat: add landing page variants and switchable home
