+++ .env.example
@@ -6,3 +6,6 @@ EXPORT_DIR=exports
 
 # Dev server port
 PORT=8000
+
+# Landing page variant to use on the home route
+NEXT_PUBLIC_LANDING=v1
